moved milestone test to own package
